pyproject.toml
[tool.poetry]
name = "plot_mover"
version = "0.1.0"
description = ""
authors = ["Maksym Leonov <maxbanton@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
PyYAML = "^5.4.1"

[tool.poetry.dev-dependencies]
flake8 = "^3.9.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
